Package reorganization: s/io.pivotal/org.cftoolsuite/g
