+++ footprints/local/support/config-server/src/main/java/org/cftoolsuite/cfapp/support/config/App.java
@@ -1,11 +1,11 @@
-package io.pivotal.cfapp.support.discovery;
+package org.cftoolsuite.cfapp.support.config;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+import org.springframework.cloud.config.server.EnableConfigServer;
 
 @SpringBootApplication
-@EnableEurekaServer
+@EnableConfigServer
 public class App {
 
 	public static void main(String[] args) {
