+++ footprints/local/support/discovery-service/src/main/java/org/cftoolsuite/cfapp/support/discovery/App.java
@@ -1,14 +1,11 @@
-package io.pivotal.cfapp.support.console;
+package org.cftoolsuite.cfapp.support.discovery;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-import de.codecentric.boot.admin.server.config.EnableAdminServer;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
-@EnableAdminServer
-@EnableDiscoveryClient
+@EnableEurekaServer
 public class App {
 
 	public static void main(String[] args) {
